% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{Arith,GeneBCMatrix,logical-method}
\alias{Arith,GeneBCMatrix,logical-method}
\title{Arithmetic operator for GeneBCMatrix}
\usage{
\S4method{Arith}{GeneBCMatrix,logical}(e1, e2)
}
\arguments{
\item{e1}{GeneBCMatrix}

\item{e2}{logical}
}
\value{
result from operator
}
\description{
Arithmetic operator for GeneBCMatrix
}
