% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{Arith,GeneBCMatrix,numeric-method}
\alias{Arith,GeneBCMatrix,numeric-method}
\title{Arithmetic operator for GeneBCMatrix}
\usage{
\S4method{Arith}{GeneBCMatrix,numeric}(e1, e2)
}
\arguments{
\item{e1}{GeneBCMatrix}

\item{e2}{numeric}
}
\value{
result from operator
}
\description{
Arithmetic operator for GeneBCMatrix
}

