% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{Arith,numeric,GeneBCMatrix-method}
\alias{Arith,numeric,GeneBCMatrix-method}
\title{Arithmetic operator for GeneBCMatrix}
\usage{
\S4method{Arith}{numeric,GeneBCMatrix}(e1, e2)
}
\arguments{
\item{e1}{numeric}

\item{e2}{GeneBCMatrix}
}
\value{
result from operator
}
\description{
Arithmetic operator for GeneBCMatrix
}
