% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\name{binary_op_right}
\alias{binary_op_right}
\title{Delegate a binary operator to the matrix slot}
\usage{
binary_op_right(e1, e2)
}
\arguments{
\item{e1}{LHS}

\item{e2}{RHS}
}
\value{
result from operator
}
\description{
Delegate a binary operator to the matrix slot
}
