% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r, R/util.r
\name{bind variables to fix data.table errors when packaging}
\alias{bind variables to fix data.table errors when packaging}
\alias{bind variables to fix data.table errors when packaging}
\title{Input/Output functions}
\description{
Input/Output functions

Utility functions
}
