% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{cell_composition}
\alias{cell_composition}
\title{Tabulate the proportion of each cell type in respective classes}
\usage{
cell_composition(clu, anno = NULL, n_digits = 5)
}
\arguments{
\item{clu}{An array of integer labels from 1 to k (length = number of cells)}

\item{anno}{Annotation for the k classes (can be an array of strings)}

\item{n_digits}{Number of digits for to display}
}
\value{
A table of the break down of the numbers of each label (with optional annotation)
}
\description{
Tabulate the proportion of each cell type in respective classes
}
\examples{
\dontrun{
 cell_composition(clu)
}
}

