% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{colMeans,GeneBCMatrix-method}
\alias{colMeans,GeneBCMatrix-method}
\title{colMeans for GeneBCMatrix objects}
\usage{
\S4method{colMeans}{GeneBCMatrix}(x, na.rm, dims = 1)
}
\arguments{
\item{x}{a GeneBCMatrix object}

\item{na.rm}{na.rm argument base function}

\item{dims}{dims argument passed to base function}
}
\value{
Result of colMeans function on exprs() on GeneBCMatrix
}
\description{
colMeans for GeneBCMatrix objects
}
\examples{
average_umis_per_cell = colMeans(gbm1)
}
