% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{colSums,GeneBCMatrix-method}
\alias{colSums,GeneBCMatrix-method}
\title{colSums for GeneBCMatrix objects}
\usage{
\S4method{colSums}{GeneBCMatrix}(x, na.rm, dims = 1)
}
\arguments{
\item{x}{a GeneBCMatrix object}

\item{na.rm}{na.rm argument base function}

\item{dims}{dims argument passed to base function}
}
\value{
Result of colSums function on exprs() on GeneBCMatrix
}
\description{
colSums for GeneBCMatrix objects
}
\examples{
umi_counts_per_cell = colSums(gbm1)
}
