% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{Compare,logical,GeneBCMatrix-method}
\alias{Compare,logical,GeneBCMatrix-method}
\title{Compare operator for GeneBCMatrix}
\usage{
\S4method{Compare}{logical,GeneBCMatrix}(e1, e2)
}
\arguments{
\item{e1}{GeneBCMatrix}

\item{e2}{GeneBCMatrix}
}
\value{
result from operator
}
\description{
Compare operator for GeneBCMatrix
}

