% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{compress_sequences}
\alias{compress_sequences}
\title{Compress a vector of sequences into 2-bit representation}
\usage{
compress_sequences(seqs)
}
\arguments{
\item{seqs}{Vector of nucleotide sequences to compress}
}
\value{
A vector of integer64 objects
}
\description{
Compress a vector of sequences into integer64 objects containing
a 2-bit representation. Ns are not allowed
}
