% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{compute_sseq_params}
\alias{compute_sseq_params}
\title{Compute sSeq parameters}
\usage{
compute_sseq_params(x, zeta_quantile = 0.995)
}
\arguments{
\item{x}{(cell x gene) count matrix}

\item{zeta_quantile}{Take this quantile of the method-of-moments dispersion estimates as the shrinkage target}
}
\description{
Compute sSeq parameters
}
