% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{concatenate_gene_bc_matrices}
\alias{concatenate_gene_bc_matrices}
\title{Concatenate gene-barcode matrices}
\usage{
concatenate_gene_bc_matrices(gbms)
}
\arguments{
\item{gbms}{a list of GeneBCMatrix objects to concatenate}
}
\value{
A new GeneBCMatrix containing all the barcodes in the given matrices
}
\description{
Concatenate gene-barcode matrices
}
\examples{
big_gbm <- concatenate_gene_bc_matrices(list(gbm1, gbm2))

}
