% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.r
\docType{data}
\name{constants}
\alias{CELL_COUNT_METRIC}
\alias{constants}
\title{Constants}
\format{An object of class \code{character} of length 1.}
\usage{
CELL_COUNT_METRIC
}
\value{
none
}
\description{
global constants
}
\keyword{datasets}

