% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{csv_value_to_numeric}
\alias{csv_value_to_numeric}
\title{Read value from metrics CSV}
\usage{
csv_value_to_numeric(x)
}
\arguments{
\item{x}{String to convert}
}
\value{
Numeric version of string
}
\description{
Read value from metrics CSV
}
