% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{dim,GeneBCMatrix-method}
\alias{dim,GeneBCMatrix-method}
\title{dim for GeneBCMatrix objects}
\usage{
\S4method{dim}{GeneBCMatrix}(x)
}
\arguments{
\item{x}{a GeneBCMatrix object}
}
\value{
Result of dim function on exprs() on GeneBCMatrix
}
\description{
dim for GeneBCMatrix objects
}
\examples{
gene_count = dim(gbm1)[1]
cell_barcode_count = dim(gbm1)[2]
}

