% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{dimnames,GeneBCMatrix-method}
\alias{dimnames,GeneBCMatrix-method}
\title{dimnames for GeneBCMatrix objects}
\usage{
\S4method{dimnames}{GeneBCMatrix}(x)
}
\arguments{
\item{x}{a GeneBCMatrix object}
}
\value{
Result of dimnames function on exprs() on GeneBCMatrix
}
\description{
dimnames for GeneBCMatrix objects
}
\examples{
gene_ids = dimnames(gbm1)[1]
cell_barcodes = dimnames(gbm1)[2]
}

