% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{directory.exists}
\alias{directory.exists}
\title{Check if a directory exists}
\usage{
directory.exists(dir_name, stop_message = NULL)
}
\arguments{
\item{path_name}{Directory name}

\item{stop}{(Optional) stop if the file does not exist}
}
\value{
Logical TRUE if exists and FALSE otherwise
}
\description{
Check if a directory exists
}
