% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{download_file_and_check}
\alias{download_file_and_check}
\title{Download file to local and check if it exists}
\usage{
download_file_and_check(url_name, loc_name)
}
\arguments{
\item{url_name}{File URL}
}
\value{
loc_name Local file path where the file is stored
}
\description{
Download file to local and check if it exists
}

