% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{equalize_gbms}
\alias{equalize_gbms}
\title{Equalize a list of gbms by subsampling}
\usage{
equalize_gbms(gbm_list)
}
\arguments{
\item{gbm_list}{A list of GeneBCMatrix objects to merge}
}
\value{
A list of GeneBCMatrix objects after subsampling
}
\description{
Equalize a list of gbms by subsampling
}
\examples{
\dontrun{
equalize_gbms(list(gbm1,gbm2))
}
}
