% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{format_barcode_string}
\alias{format_barcode_string}
\title{Format a barcode string}
\usage{
format_barcode_string(barcode_seq, gem_group)
}
\arguments{
\item{barcode_seq}{Cell barcode sequence}

\item{gem_group}{Gem group number}
}
\value{
A formatted barcode string
}
\description{
Format a barcode string
}
