% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gbm1}
\alias{gbm1}
\title{An example  GeneBarcodeMatrix object containing 7827 cells and 35635 genes.}
\format{An object of class \code{GeneBCMatrix} with 35635 rows and 50 columns.}
\usage{
gbm1
}
\value{
none. data documentation.
}
\description{
Data contains peripheral blood mononuclear cells (PBMCs) from human individual 1.
}
\keyword{datasets}
