% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gbm2}
\alias{gbm2}
\title{An example  GeneBarcodeMatrix object containing 9515 cells and 35635 genes.}
\format{An object of class \code{GeneBCMatrix} with 35635 rows and 65 columns.}
\usage{
gbm2
}
\value{
none. data documentation.
}
\description{
Data contains peripheral blood mononuclear cells (PBMCs) from human individual 2.
}
\keyword{datasets}

