% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gbm_pheatmap}
\alias{gbm_pheatmap}
\title{Plot heatmap with specified gene and cell ordering}
\usage{
gbm_pheatmap(gbm, genes_to_plot, cells_to_plot, n_genes = 5, colour = NULL,
  limits = c(-3, 3))
}
\arguments{
\item{gbm}{A GeneBCMatrix object (recommended to be normalized by UMI counts and log-transformed)}

\item{genes_to_plot}{A list of gene symbols (or output from function: prioritize_genes)}

\item{cells_to_plot}{A dataframe with cell names and cell indices (or output from function: order_cell_by_clusters)}

\item{n_genes}{Number of genes to include in the heatmap}

\item{colour}{Customized colors for each class label on the colour axes}

\item{limits}{(min,max) saturates values on the color bar}
}
\value{
none. a pheatmap is generated.
}
\description{
Generate a ggplot object that highlights specified class labels
}
\examples{
\dontrun{
gbm_pheatmap(gbm,genes_to_plot,cells_to_plot)
}

}
