% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_barcode_sequence}
\alias{get_barcode_sequence}
\title{Extract the cell barcode sequence from one or more barcode strings}
\usage{
get_barcode_sequence(barcode_str)
}
\arguments{
\item{barcode_str}{Formatted cell barcode sequence}
}
\value{
A cell barcode sequence string
}
\description{
Extract the cell barcode sequence from one or more barcode strings
}

