% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{get_correletion_scores}
\alias{get_correletion_scores}
\title{Compute the correlation score with a bulk population profile}
\usage{
get_correletion_scores(gbm_in, compare_exprs, compare_gene_ids,
  method = "spearman")
}
\arguments{
\item{gbm_in}{A GeneBCMatrix (recommended to be normalized by UMI counts)}

\item{compare_exprs}{A matrix that includes expressions with one typical expression per row}

\item{compare_gene_ids}{The gene ids corresponding to the gene expression}
}
\value{
A matrix of correlation scores (each row is a cell in gbm_in) based on the intersection of genes between gbm_in and compare_exprs
}
\description{
Compute the correlation score with a bulk population profile
}
\examples{
\dontrun{
 get_correletion_scores(gbm, bulk_profile$avg_expr, bulk_profile$genes$id)
}
}
