% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{get_custom_col}
\alias{get_custom_col}
\title{Get custmoized colours}
\usage{
get_custom_col(n, style = "default")
}
\arguments{
\item{n}{Number of colors needed}

\item{style}{The color palette}
}
\value{
a character array of up to 13 colors
}
\description{
Generate an array of up to 13 colors from custmoized color palette
}
\examples{
get_custom_col(13)

}

