% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_gem_group}
\alias{get_gem_group}
\title{Extract the gem group from one or more barcode strings}
\usage{
get_gem_group(barcode_str)
}
\arguments{
\item{barcode_str}{Formatted cell barcode sequence}
}
\value{
A gem group integer
}
\description{
Extract the gem group from one or more barcode strings
}

