% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{get_gene_dispersion}
\alias{get_gene_dispersion}
\title{Compute the normalized dispersion for each gene}
\usage{
get_gene_dispersion(gbm_in, num_bins = 20)
}
\arguments{
\item{num_bins}{The number of bins to calculate the dispersion norm}

\item{gbm_norm}{A GeneBCMatrix normalized by UMI counts (or output from normalize_barcode_sums_to_median())}
}
\value{
A data frame containing \itemize{
\item "mean" - The mean expression each gene
\item "var" - The variance of each gene
\item "dispersion" - The computed dispersion (var/mean) of each gene
\item "dispersion_norm" - The dispersion noramlized by binning
\item "id" - gene ids
\item "symbol" - gene symbols
}
}
\description{
Compute the normalized dispersion for each gene
}
\examples{
\dontrun{
 get_gene_dispersion(gbm1)
}
}

