% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_gene_index}
\alias{get_gene_index}
\title{Get a gene index from a valid gene symbol}
\usage{
get_gene_index(gbm, gene_symbol)
}
\arguments{
\item{gbm}{A GeneBCMatrix to get the non-zero genes from}

\item{gene_symbol}{A gene symbol string}
}
\value{
The indices from a gene index
}
\description{
Get a gene index from a valid gene symbol
}
