% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{get_genome_in_matrix_path}
\alias{get_genome_in_matrix_path}
\title{Get list of genomes in Cell Ranger output}
\usage{
get_genome_in_matrix_path(matrix_path, genome = NULL)
}
\arguments{
\item{matrix_path}{Path to the output directory produced by the Cell Ranger pipeline}

\item{barcode_filtered}{Path to filtered matrix if true, otherwise raw matrix}
}
\value{
A character vector of the genomes found
}
\description{
Get list of genomes in Cell Ranger output
}

