% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_genomes_in_matrix_dir}
\alias{get_genomes_in_matrix_dir}
\title{Get list of genomes in Cell Ranger output}
\usage{
get_genomes_in_matrix_dir(pipestance_path, barcode_filtered)
}
\arguments{
\item{pipestance_path}{Path to the output directory produced by the Cell Ranger pipeline}

\item{barcode_filtered}{Path to filtered matrix if true, otherwise raw matrix}
}
\value{
A character vector of the genomes found
}
\description{
Get list of genomes in Cell Ranger output
}

