% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{get_matrix_dir_path}
\alias{get_matrix_dir_path}
\title{Get path to matrices in Cell Ranger output}
\usage{
get_matrix_dir_path(pipestance_path, barcode_filtered)
}
\arguments{
\item{pipestance_path}{Path to the output directory produced by the Cell Ranger pipeline}

\item{barcode_filtered}{If true, get the path to the barcode filtered matrices; else to the raw filtered matrices}
}
\value{
The path to the matrices directory
}
\description{
Get path to matrices in Cell Ranger output
}
