% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{get_matrix_from_h5}
\alias{get_matrix_from_h5}
\title{Load gene-barcode matrix from a specific H5 file}
\usage{
get_matrix_from_h5(filename, genome = NULL)
}
\arguments{
\item{genome}{The desired genome (e.g., 'hg19' or 'mm10')}

\item{Path}{to the gene-barcode matrix H5 file (NOT the pipestance path)}
}
\value{
A GeneBCMatrix sparse matrix where rows are genes and columns are cell-barcodes
}
\description{
Load gene-barcode matrix from a specific H5 file
}
\examples{
\dontrun{
# Load hg19 from a filtered matrix h5
gene_bc_matrix <- get_matrix_from_h5("/home/user/cellranger_output/outs/filtered_gene_bc_matrices_h5.h5", "hg19")
}
}

