% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_matrix_path}
\alias{get_matrix_path}
\title{Get path to a genome matrix in Cell Ranger output}
\usage{
get_matrix_path(pipestance_path, barcode_filtered, genome)
}
\arguments{
\item{pipestance_path}{Path to the output directory produced by the Cell Ranger pipeline}

\item{barcode_filtered}{If true, get the path to the barcode filtered matrix; else to the raw filtered matrix}

\item{genome}{The desired genome (e.g., 'hg19' or 'mm10')}
}
\value{
The path to a matrix directory
}
\description{
Get path to a genome matrix in Cell Ranger output
}

