% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_mean_mapped_reads_per_cell}
\alias{get_mean_mapped_reads_per_cell}
\title{Get mean mapped reads per cell}
\usage{
get_mean_mapped_reads_per_cell(gbm)
}
\arguments{
\item{gbm}{A GeneBCMatrix object}
}
\value{
Mean confidently mapped, barcoded reads per cell
}
\description{
Get mean mapped reads per cell
}
\examples{
\dontrun{
mapped_rpc <- get_mean_mapped_reads_per_cell(gbm)
}
}
