% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_mean_raw_reads_per_cell}
\alias{get_mean_raw_reads_per_cell}
\title{Get mean raw reads per cell}
\usage{
get_mean_raw_reads_per_cell(gbm)
}
\arguments{
\item{gbm}{A GeneBCMatrix object}
}
\value{
Mean raw reads per cell
}
\description{
Get mean raw reads per cell
}
\examples{
\dontrun{
raw_rpc <- get_mean_raw_reads_per_cell(gbm1)
}
}

