% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_metric}
\alias{get_metric}
\title{Get a metric from the metric summary}
\usage{
get_metric(gbm, metric_name)
}
\arguments{
\item{gbm}{A GeneBCMatrix object}

\item{metric_name}{Metric name}
}
\value{
numeric metric value
}
\description{
Get a metric from the metric summary
}

