% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_nonzero_genes}
\alias{get_nonzero_genes}
\title{Get the non-zero gene indices from a GeneBCMatrix}
\usage{
get_nonzero_genes(gbm)
}
\arguments{
\item{gbm}{a GeneBCMatrix to get the non-zero genes from}
}
\value{
The gene (row) indices for genes with at least 1 count in the matrix
}
\description{
Get the non-zero gene indices from a GeneBCMatrix
}
\examples{
non_zero_gene_ids <- get_nonzero_genes(gbm1)

}

