% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{get_ordered_gene_list_sseq}
\alias{get_ordered_gene_list_sseq}
\title{Get ordered gene list by the sSeq differential expression method}
\usage{
get_ordered_gene_list_sseq(gbm, cluster_labels, test_ID, sseq_params, t_mat)
}
\arguments{
\item{gbm}{A GeneBCMatrix}

\item{cluster_labels}{An array of integer labels from 1 to k (length = number of cells)}

\item{test_ID}{A label between 1 to k to be tested}

\item{sseq_params}{List returned by compute_sseq_params}

\item{t_mat}{BC x gene matrix}
}
\value{
an array of genes sorted by the mean differential expression between testID and the rest
}
\description{
Get ordered gene list by the sSeq differential expression method
}

