% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{get_ordered_gene_list_unique}
\alias{get_ordered_gene_list_unique}
\title{Get ordered gene list by gene uniqueness}
\usage{
get_ordered_gene_list_unique(gbm, cluster_labels, test_ID)
}
\arguments{
\item{gbm}{A GeneBCMatrix}

\item{cluster_labels}{An array of integer labels from 1 to k (length = number of cells)}

\item{test_ID}{A label between 1 to k to be tested}
}
\value{
an array of genes sorted by the mean differential expression between testID and the rest
}
\description{
Get ordered gene list by gene uniqueness
}
