% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_read_count_mtx}
\alias{get_read_count_mtx}
\title{Get the read count matrix}
\usage{
get_read_count_mtx(gbm)
}
\arguments{
\item{gbm}{A GeneBCMatrix object}
}
\value{
A dgTMatrix with the same dimension as exprs(gbm)
}
\description{
Get the read count matrix
}
\examples{
\dontrun{
 get_read_count_mtx(gbm1)
 }
}
