% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_summary_csv_path}
\alias{get_summary_csv_path}
\title{Get summary csv in Cell Ranger output}
\usage{
get_summary_csv_path(pipestance_path)
}
\arguments{
\item{pipestance_path}{Path to the output directory produced by the Cell Ranger pipeline}
}
\value{
Path to the summary CSV
}
\description{
Get summary csv in Cell Ranger output
}

