% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{get_tsne_path}
\alias{get_tsne_path}
\title{Load Cell Ranger t-SNE results from a specified analysis folder}
\usage{
get_tsne_path(analysis_path)
}
\arguments{
\item{analysis_path}{Path to the analysis output directory produced by Cell Ranger}
}
\value{
Path to t-SNE results
}
\description{
Load Cell Ranger t-SNE results from a specified analysis folder
}
\examples{
\dontrun{
tsne_path <- get_tsne_path(analysis_path)
}
}

