% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{load_analysis_results_from_path}
\alias{load_analysis_results_from_path}
\title{Load Cell Ranger secondary analysis results from a specified analysis folder}
\usage{
load_analysis_results_from_path(analysis_path)
}
\arguments{
\item{analysis_path}{Path to the analysis output directory produced by Cell Ranger}
}
\value{
A list containing: \itemize{
\item pca - pca projection
\item tsne - tsne projection
\item clustering - clustering results for various clustering methods
}
}
\description{
Load Cell Ranger secondary analysis results from a specified analysis folder
}
\examples{
\dontrun{
analysis <- load_analysis_results_from_path("/home/user/cellranger_output/outs/analysis")
}
}
