% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{load_cellranger_matrix_from_files}
\alias{load_cellranger_matrix_from_files}
\title{Load data from the 10x Genomics Cell Ranger pipeline by specifying individual file paths}
\usage{
load_cellranger_matrix_from_files(mat_fn, gene_fn, barcode_fn,
  summary_fn = NULL)
}
\arguments{
\item{mat_fn}{Path to a .mtx file including matirx entries}

\item{gene_fn}{Path to a .tsv file including gene information}

\item{barcode_fn}{Path to a .tsv file including cell barcode information}

\item{summary_fn}{(Optional) Path to a .csv file including cell barcode information}
}
\value{
A GeneBCMatrix sparse matrix where rows are genes and columns are cell-barcodes
}
\description{
Load data from the 10x Genomics Cell Ranger pipeline by specifying individual file paths
}
\examples{
\dontrun{
gbm <- load_cellranger_matrix_from_files("matrix.mtx","genes.tsv","barcodes.tsv")
}
}
