% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{load_cellranger_summary}
\alias{load_cellranger_summary}
\title{Load metrics summary from the Cell Ranger pipeline}
\usage{
load_cellranger_summary(pipestance_path)
}
\arguments{
\item{pipestance_path}{Path to the output directory produced by the Cell Ranger pipeline}
}
\value{
The summary csv in a data frame
}
\description{
Load metrics summary from the Cell Ranger pipeline
}
\examples{
\dontrun{
single_cell_metrics <- load_cellranger_summary("/home/user/cellranger_output")
}
}

