% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{load_molecule_info}
\alias{load_molecule_info}
\title{Load molecule info h5 file from the Cell Ranger pipeline}
\usage{
load_molecule_info(gbm)
}
\arguments{
\item{gbm}{GeneBCMatrix object to load into}

\item{barcode_len}{Length of cell barcode in nucleotides}
}
\value{
A GeneBCMatrix object containing the loaded molecule info
}
\description{
Load molecule info h5 file from the Cell Ranger pipeline
}
\examples{
\dontrun{
new_reads_per_cell = 10000 # downsampling read depth per cell to 10000
gbm1 = load_molecule_info(gbm1)
gbm1_subsampled = subsample_gene_bc_matrix(gbm1, 10000)

# Note molecule info stored as gbm1@molecule_info
}
}

