% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{load_molecule_info_from_path}
\alias{load_molecule_info_from_path}
\title{Load molecule info h5 file from a specified .h5 file}
\usage{
load_molecule_info_from_path(gbm, f_path)
}
\arguments{
\item{gbm}{GeneBCMatrix object to load into}

\item{f_path}{Path to the molecule_info.h5}
}
\value{
A GeneBCMatrix object containing the loaded molecule info
}
\description{
Load molecule info h5 file from a specified .h5 file
}
\examples{
\dontrun{
gbm1 = load_molecule_info(gbm1, f_path)
# Note molecule info stored as gbm1@molecule_info
}
}
