% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{log_gene_bc_matrix}
\alias{log_gene_bc_matrix}
\title{Take the element-wise log(1+count) of a gene-barcode matrix while maintaining sparsity}
\usage{
log_gene_bc_matrix(gbm, base = 2)
}
\arguments{
\item{gbm}{a GeneBCMatrix to log}

\item{base}{base of logarithm to use}
}
\value{
A GeneBCMatrix with logged counts
}
\description{
Take the element-wise log(1+count) of a gene-barcode matrix while maintaining sparsity
}
\examples{
gbm_log <- log_gene_bc_matrix(gbm1)

}

