% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{Logic,GeneBCMatrix,numeric-method}
\alias{Logic,GeneBCMatrix,numeric-method}
\title{Logical operator for GeneBCMatrix}
\usage{
\S4method{Logic}{GeneBCMatrix,numeric}(e1, e2)
}
\arguments{
\item{e1}{GeneBCMatrix}

\item{e2}{GeneBCMatrix}
}
\value{
result from operator
}
\description{
Logical operator for GeneBCMatrix
}
