% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{Math,GeneBCMatrix-method}
\alias{Math,GeneBCMatrix-method}
\title{Math operator for GeneBCMatrix}
\usage{
\S4method{Math}{GeneBCMatrix}(x)
}
\arguments{
\item{x}{GeneBCMatrix}
}
\value{
result from operator
}
\description{
Math operator for GeneBCMatrix
}
