% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{Math2,GeneBCMatrix-method}
\alias{Math2,GeneBCMatrix-method}
\title{Math2 operator for GeneBCMatrix}
\usage{
\S4method{Math2}{GeneBCMatrix}(x)
}
\arguments{
\item{x}{GeneBCMatrix}
}
\value{
result from operator
}
\description{
Math2 operator for GeneBCMatrix
}

