% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge clusters among cluster assignments}
\usage{
merge_clusters(cluster_result, cluster_select, colour = NULL)
}
\arguments{
\item{cluster_result}{A vector of integer class labels for each cell}

\item{cluster_select}{A vector of cluster (integer) ids to select}

\item{colour}{(Optional) colours assignments for original clusters}
}
\value{
A list containing: \itemize{
\item ids - new (consecutive) id labels for each cell after merging integer labels
\item col - new colours (mapping to the ids) corresponding to the integer labels if colour is specified, otherwise equals NULL
}
}
\description{
Merge clusters among cluster assignments
}
\examples{
merge_clusters_test(rep(1:5,4),c(1,3,4))
}

