% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{nb_exact_test}
\alias{nb_exact_test}
\title{Negative binomial exact test as in sSeq}
\usage{
nb_exact_test(x_a, x_b, s_a, s_b, mu, phi)
}
\arguments{
\item{x_a}{Total count for a single gene in group A}

\item{x_b}{Total count for the same gene in group B}

\item{s_a}{Size factor for group A}

\item{s_b}{Size factor for group B}

\item{mu}{Common mean for gene}

\item{phi}{Common dispersion for gene}
}
\description{
Negative binomial exact test as in sSeq
}
