% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{ncol,GeneBCMatrix-method}
\alias{ncol,GeneBCMatrix-method}
\title{ncol for GeneBCMatrix objects}
\usage{
\S4method{ncol}{GeneBCMatrix}(x)
}
\arguments{
\item{x}{a GeneBCMatrix object}
}
\value{
Result of ncol function on exprs() on GeneBCMatrix
}
\description{
ncol for GeneBCMatrix objects
}
\examples{
cell_barcode_count = ncol(gbm1)
}
