% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\name{newGeneBCMatrix}
\alias{newGeneBCMatrix}
\title{Create a new GeneBCMatrix}
\usage{
newGeneBCMatrix(mat, fd, pd, template = NULL)
}
\arguments{
\item{mat}{A sparse matrix whose rows are genes and columns are cell-barcodes}

\item{fd}{A dataframe with gene IDs as row names and additional metadata such as gene symbols as columns.}

\item{pd}{A dataframe with cell barcodes as row names and any additional metadata columns.}

\item{template}{A GeneBCMatrix to copy slots from}
}
\value{
A new GeneBCMatrix object
}
\description{
Create a new GeneBCMatrix
}
\examples{
\dontrun{
matrix = exprs(gbm1)
featureData = fData(gbm1)
phenoData = pData(gbm1)

# Note including a template copies all other slots of gbm to new object
new_gbm = newGeneBCMatrix(matrix, fd=featureData, pd=phenoData, template=gbm1)
}
}
