% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize_barcode_sums_to_median}
\alias{normalize_barcode_sums_to_median}
\title{Normalization functions}
\usage{
normalize_barcode_sums_to_median(gbm)
}
\arguments{
\item{gbm}{a GeneBCMatrix to normalize}
}
\value{
A GeneBCMatrix with barcode-normalized values
}
\description{
Normalize barcodes by sum to the median barcode by sum
}
\examples{
gbm_norm_by_bc <- normalize_barcode_sums_to_median(gbm1)

}
