% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{nrow,GeneBCMatrix-method}
\alias{nrow,GeneBCMatrix-method}
\title{nrow for GeneBCMatrix objects}
\usage{
\S4method{nrow}{GeneBCMatrix}(x)
}
\arguments{
\item{x}{a GeneBCMatrix object}
}
\value{
Result of nrow function on exprs() on GeneBCMatrix
}
\description{
nrow for GeneBCMatrix objects
}
\examples{
gene_count = nrow(gbm1)
}

