% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{order_cell_by_clusters}
\alias{order_cell_by_clusters}
\title{Order cells based on clusters}
\usage{
order_cell_by_clusters(gbm, clu)
}
\arguments{
\item{gbm}{A GeneBCMatrix}

\item{clu}{An array of integer labels from 1 to k (length = number of cells)}
}
\value{
A list of (n_top_genes * number of clusters) genes (may include duplicates)
}
\description{
Order cells based on clusters
}
\examples{
\dontrun{
 order_cell_by_clusters(gbm_log,clu)
}
}

