% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_barcode_counts}
\alias{plot_barcode_counts}
\title{Plotting functions}
\usage{
plot_barcode_counts(gene_bc_matrix)
}
\arguments{
\item{gene_bc_matrix}{A GeneBCMatrix object}
}
\value{
none. plot is generated.
}
\description{
Plot barcode distribution
}
\details{
Plot the barcode counts in descending order on a log-scale
}
\examples{
\dontrun{
plot_barcode_counts(gbm1)
}

}
