% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_population_corr}
\alias{plot_population_corr}
\title{Plot correlation heatmap}
\usage{
plot_population_corr(pop_avg, method = "spearman")
}
\arguments{
\item{pop_avg}{An input matrix (rownames will be used) where rows are samples and columns are features}

\item{method}{The correlation method for the cor() function to compute pairwise correlations}
}
\value{
none. a pheatmap is generated.
}
\description{
Generate a pheatmap that shows the correlation between pairs of rows in an input matrix
}
\examples{
plot_population_corr(gbm,genes_to_plot,cells_to_plot)

}

