% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{+,GeneBCMatrix,missing-method}
\alias{+,GeneBCMatrix,missing-method}
\title{Unary identity of GeneBCMatrix objects}
\usage{
\S4method{+}{GeneBCMatrix,missing}(e1, e2)
}
\arguments{
\item{e1}{GeneBCMatrix}

\item{e2}{dummy}
}
\value{
result from operator
}
\description{
Unary identity of GeneBCMatrix objects
}

