% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{prioritize_genes}
\alias{prioritize_genes}
\title{Prioritize genes that are most significant markers for each cluster}
\usage{
prioritize_genes(gbm, clu, method)
}
\arguments{
\item{gbm}{A GeneBCMatrix (recommended to be normalized by UMI counts and log-transformed)}

\item{clu}{An array of integer labels from 1 to k (length = number of cells)}

\item{method}{Method to priortize genes: currently only supports mean difference}

\item{n_top_genes}{Number of top genes for each cluster group to consider}
}
\description{
Prioritize genes that are most significant markers for each cluster
}
\examples{
\dontrun{
 prioritize_top_genes(gbm_log,clu,3,'mean_diff')
}
}

