% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{prioritize_top_genes}
\alias{prioritize_top_genes}
\title{Prioritize genes that are most significant markers for each cluster}
\usage{
prioritize_top_genes(gbm, clu, method, logscale = TRUE, p_cutoff = 0.05,
  min_mean = 1, min_log2fc = 0, order_by = "pvalue")
}
\arguments{
\item{gbm}{A GeneBCMatrix (recommended to be normalized by UMI counts and log-transformed for the  "mean_diff" method). For the "sseq" method, this should be a raw (unnormalized) GeneBCMatrix.}

\item{clu}{An array of integer labels from 1 to k (length = number of cells)}

\item{method}{Method to prioritize genes: currently only supports mean difference ('mean_diff', 'unique', or 'sseq')}

\item{logscale}{Logical if the input is log-scale or not (default: TRUE)}

\item{p_cutoff}{If method is "sseq," only consider genes w/ adjusted p-value below this value}

\item{min_mean}{If method is "sseq," only consider genes w/ mean normalized UMI counts/cell exceeding this value}

\item{min_log2fc}{If method is "sseq," only consider genes with at least this log2 fold-change}

\item{order_by}{If method is "sseq," sort genes by 'pvalue' or by 'log2fc'}

\item{n_top_genes}{Number of top genes for each cluster group to consider}
}
\description{
Prioritize genes that are most significant markers for each cluster
}
\examples{
\dontrun{
 prioritize_top_genes(gbm, cluster_labels, 'sseq')
}
}

