% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{rowMeans,GeneBCMatrix-method}
\alias{rowMeans,GeneBCMatrix-method}
\title{rowMeans for GeneBCMatrix objects}
\usage{
\S4method{rowMeans}{GeneBCMatrix}(x, na.rm, dims = 1)
}
\arguments{
\item{x}{a GeneBCMatrix object}

\item{na.rm}{na.rm argument base function}

\item{dims}{dims argument passed to base function}
}
\value{
Result of rowMeans function on exprs() on GeneBCMatrix
}
\description{
rowMeans for GeneBCMatrix objects
}
\examples{
average_umis_per_gene = rowMeans(gbm1)
}
