% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\docType{methods}
\name{rowSums,GeneBCMatrix-method}
\alias{rowSums,GeneBCMatrix-method}
\title{rowSums for GeneBCMatrix objects}
\usage{
\S4method{rowSums}{GeneBCMatrix}(x, na.rm, dims = 1)
}
\arguments{
\item{x}{a GeneBCMatrix object}

\item{na.rm}{na.rm argument base function}

\item{dims}{dims argument passed to base function}
}
\value{
Result of rowSums function on exprs() on GeneBCMatrix
}
\description{
rowSums for GeneBCMatrix objects
}
\examples{
umi_counts_per_gene = rowSums(gbm1)
}

