% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{run_differential_expression}
\alias{run_differential_expression}
\title{Run differential expression analysis between two sets of cells}
\usage{
run_differential_expression(gbm, cell_indices0, cell_indices1)
}
\arguments{
\item{gbm}{GeneBCMatrix}

\item{cell_indices0}{(integer) Indices of cells in group A}

\item{cell_indices1}{(integer) Indices of cells in group B}
}
\value{
A data frame containing differential expression results for each gene.
}
\description{
Run differential expression analysis between two sets of cells
}
\examples{
\dontrun{
de_result <- run_differential_expression(gbm, 1:500, 501:1000)
top_results <- de_result \%>\% filter(p_adj < 0.05) \%>\% arrange(desc(log2fc))
head(top_results)
}
}

