% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{run_kmeans_clustering}
\alias{run_kmeans_clustering}
\title{Run k-means clustering}
\usage{
run_kmeans_clustering(pca, k, n_pcs = NULL, ...)
}
\arguments{
\item{pca}{A PCA result from run_pca()}

\item{k}{number of clusters}

\item{n_pcs}{The number of top pca components to use (default is to use all pcs)}

\item{...}{parameters to pass to kmeans}
}
\value{
See ?kmeans. The item of interest is the 'cluster' element
}
\description{
Run k-means clustering
}
\examples{
pca <- run_pca(gbm1, n_pcs=10)
km_res <- run_kmeans_clustering(pca, k=5)

}
\seealso{
kmeans
}

