% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{save_cellranger_matrix_h5}
\alias{save_cellranger_matrix_h5}
\title{Save a single-genome GeneBCMatrix as an h5 file for input into `cellranger reanalyze`}
\usage{
save_cellranger_matrix_h5(gbm, out_filename, genome_name, template = NULL,
  overwrite = FALSE)
}
\arguments{
\item{gbm}{GeneBCMatrix to write}

\item{out_filename}{The HDF5 filename to write to}

\item{genome_name}{The original genome name (e.g., GRCh38)}

\item{template}{Path to the original matrix.h5 file.}

\item{overwrite}{Overwrite the HDF5 file if it exists}
}
\description{
Save a single-genome GeneBCMatrix as an h5 file for input into `cellranger reanalyze`
}
\examples{
\dontrun{
# Save a gbm as a matrix HDF5 file
save_cellranger_matrix_h5(gbm, "new_matrix.h5", "GRCh38")
}
}
