% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{select_cells_by_clusterID}
\alias{select_cells_by_clusterID}
\title{Select cells according to cluster IDs}
\usage{
select_cells_by_clusterID(cluster_result, cluster_select)
}
\arguments{
\item{cluster_result}{A vector of class labels for each cell}

\item{cluster_select}{A vector (or scalar) of cluster ids to select}
}
\value{
Logical whether or not a cell is selected
}
\description{
Select cells according to cluster IDs
}
\examples{
select_cells_by_clusterID(rep(1:5,4),c(1,3,4))
}
