% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{select_clusters_by_gene_markers}
\alias{select_clusters_by_gene_markers}
\title{Select cells by clusters according to the fraction of cells that express at least one of the input gene markers}
\usage{
select_clusters_by_gene_markers(gbm, gene_symbols, cluster, frac)
}
\arguments{
\item{gbm}{A GeneBCMatrix (recommended to be normalized by UMI counts and log-transformed)}

\item{gene_symbols}{A list of gene symbols}

\item{cluster}{A vector of class labels for each cell}

\item{frac}{Minimum fraction of cells in a cluster required for the cluster to be selected}
}
\value{
Logical whether or not a cell is selected
}
\description{
Select cells by clusters according to the fraction of cells that express at least one of the input gene markers
}
\examples{
select_cells_by_clusterID(rep(1:5,4),c(1,3,4))
}

