% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{sparse_pca}
\alias{sparse_pca}
\title{Run a fast PCA algorithm while maintaining sparsity}
\usage{
sparse_pca(x, n_pcs, mu = NULL, s = NULL, center_scale = TRUE)
}
\arguments{
\item{x}{a sparse matrix to run PCA on}

\item{n_pcs}{number of principal components to compute}

\item{mu}{column means}

\item{s}{column standard deviations}

\item{center_scale}{perform centering and scaling}
}
\value{
a list containing the rotated matrix ('x') and the principal component vectors ('rotation')
}
\description{
Run a fast PCA algorithm while maintaining sparsity
}
\examples{
\dontrun{
pca <- sparse_pca(x, n_pcs=10)
plot(pca$x[,1], pca$x[,2], xlab='PC1', ylab='PC2')
}
}

