% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{sseq_differential_expression}
\alias{sseq_differential_expression}
\title{Compute differential expression using the sSeq method}
\usage{
sseq_differential_expression(x, cond0, cond1, sseq_params, gene_ids,
  gene_symbols)
}
\arguments{
\item{x}{(cell x gene) Count matrix}

\item{cond0}{(integer) Indices of cells in group A}

\item{cond1}{(integer) Indices of cells in group B}

\item{sseq_params}{Params precomputed from whole sample}
}
\description{
Compute differential expression using the sSeq method
}
