% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{subsample_gene_bc_matrix}
\alias{subsample_gene_bc_matrix}
\title{Subsample a matrix to a given sequencing depth}
\usage{
subsample_gene_bc_matrix(gbm, target_reads_per_cell,
  target_type = "raw_reads", verbose = TRUE)
}
\arguments{
\item{gbm}{GeneBCMatrix object to subsample}

\item{target_reads_per_cell}{Target depth specified as reads per cell}

\item{target_type}{Type of reads specified in target_reads_per_cell; one of: \itemize{
\item \code{"raw_reads"} - Raw reads per cell.
\item \code{"mapped_reads"} Confidently mapped, barcoded reads per cell. Raw reads are sufficient in most cases, but if the valid barcode fraction or confidently mapped read fraction vary greatly between samples, this option should be used.
}}

\item{verbose}{Print status messages along the way}
}
\value{
A new GeneBCMatrix object that has been subsampled to the specified read depth
}
\description{
Subsample a matrix to a given sequencing depth so it is comparable to matrices from other samples
}
\examples{
\dontrun{
new_reads_per_cell = 10000 # downsampling read depth per cell to 10000
gbm1 = load_molecule_info(gbm1)
gbm1_subsampled = subsample_gene_bc_matrix(gbm1, 10000)
}
}

