% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{train_multinomial}
\alias{train_multinomial}
\title{Train a multinomial distribution from a vector}
\usage{
train_multinomial(x)
}
\arguments{
\item{x}{An input vector of values}
}
\value{
A scaled vector of frequencies
}
\description{
Train a multinomial distribution from a vector
}
\examples{
train_multinomial(1:10)
}
