% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{trunc_gbm_by_genes}
\alias{trunc_gbm_by_genes}
\title{Truncate a GeneBCMatrix according to an array of gene symbols}
\usage{
trunc_gbm_by_genes(gbm, gene_symbols)
}
\arguments{
\item{gbm}{A GeneBCMatrix}

\item{gene_symbols}{A list of gene symbols}
}
\value{
A GeneBCMatrix truncated to genes with specific gene symbols
}
\description{
Truncate a GeneBCMatrix according to an array of gene symbols
}
\examples{
gbm_trunc <- trunc_gbm_by_genes(gbm1,c('CD79A','NKG7'))

}
