% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\name{unary_arith_op}
\alias{unary_arith_op}
\title{Delegate a unary arithmetic operator to the matrix slot}
\usage{
unary_arith_op(e1, e2)
}
\arguments{
\item{e1}{GeneBCMatrix object}

\item{e2}{dummy}
}
\value{
result from operator
}
\description{
Delegate a unary arithmetic operator to the matrix slot
}
