% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gene_bc_matrix.r
\name{unary_op}
\alias{unary_op}
\title{Delegate a unary operator to the matrix slot}
\usage{
unary_op(x)
}
\arguments{
\item{x}{GeneBCMatrix object}
}
\value{
result from operator
}
\description{
Delegate a unary operator to the matrix slot
}

