% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{visualize_clusters}
\alias{visualize_clusters}
\title{Visualize class labels under a 2D projection}
\usage{
visualize_clusters(cluster_result, projection, colour = NULL, alpha = 1,
  marker_size = 0.1, title = NULL)
}
\arguments{
\item{cluster_result}{A vector or matrix of class labels for each cell (row)}

\item{projection}{A two column matrix projection of each cell}

\item{colour}{Customized colors for each class label}
}
\value{
A ggplot object with facets corresponding to each gene symbol
}
\description{
Generate a ggplot object that highlights specified class labels
}
\examples{
\dontrun{
visualize_clusters(cluster_result,tsne_proj)
}

}

