% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{visualize_gene_markers}
\alias{visualize_gene_markers}
\title{Visualize gene markers under a 2D projection}
\usage{
visualize_gene_markers(gbm, gene_probes, projection, limits = c(0, 10),
  marker_size = 0.1, title = NULL)
}
\arguments{
\item{gbm}{A GeneBCMatrix object (recommended to be normalized by UMI counts and log-transformed)}

\item{gene_probes}{A vector of gene symbols to highlight}

\item{projection}{A two column matrix projection of each cell}

\item{limits}{(min,max) saturates values on the color bar}
}
\value{
A ggplot object with facets corresponding to each gene symbol
}
\description{
Generate a ggplot object that highlights gene markers
}
\examples{
\dontrun{
visualize_gene_markers(gbm,c('gene1','gene2'),tsne_proj)
}

}

