% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{visualize_variable_genes}
\alias{visualize_variable_genes}
\title{Visualize variable genes}
\usage{
visualize_variable_genes(gene_dispersion, used = FALSE, marker_size = 1)
}
\arguments{
\item{gene_dispersion}{Dispersion for each gene (output from function get_gene_dispersion())}

\item{used}{(Optional) if TRUE will plot the used genes}

\item{marker_size}{(Optional) marker size (default: 0)}
}
\value{
none. A ggplot object that shows normalized dispersion v.s. mean expression
}
\description{
Visualize variable genes
}
\examples{
plot_population_corr(gbm,genes_to_plot,cells_to_plot)

}
