% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{download_sample}
\alias{download_sample}
\title{Download a sample and relevant information from software.10xgenomics.com}
\usage{
download_sample(sample_name, sample_dir, host, lite = TRUE)
}
\arguments{
\item{sample_name}{A name of the sample to be downloaded}

\item{sample_dir}{Directory where the data is saved}

\item{host}{The prefix of the url where the data is downloaded from}

\item{lite}{If TRUE, then molecule_info will not be downloaded}
}
\value{
None The sample_dir can be used to create a gbm object
}
\description{
Download a sample and relevant information from software.10xgenomics.com
}
\examples{
\dontrun{
 download_sample('pbmc3k','/mnt/home/jason/public_html/Download','https://s3-us-west-2.amazonaws.com/10x.files/samples/cell/')
 }
}

