% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{load_cellranger_analysis_results}
\alias{load_cellranger_analysis_results}
\title{Load Cell Ranger secondary analysis results}
\usage{
load_cellranger_analysis_results(pipestance_path)
}
\arguments{
\item{pipestance_path}{Path to the output directory produced by Cell Ranger}
}
\value{
A list containing: \itemize{
\item pca - pca projection
\item tsne - tsne projection
\item kmeans - kmeans results for various values of K
}
}
\description{
Load Cell Ranger secondary analysis results
}
\examples{
\dontrun{
analysis <- load_cellranger_analysis_results("/home/user/cellranger_output")
# Plot t-SNE projection
plot(analysis$tsne$TSNE.1, analysis$tsne$TSNE.2)

# Color by kmeans with K=4
plot(analysis$tsne$TSNE.1, analysis$tsne$TSNE.2, col=analysis$kmeans[["2_clusters"]]$Cluster)
}
}

