% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{load_molecule_info}
\alias{load_molecule_info}
\title{Load molecule info h5 file from the Cell Ranger pipeline}
\usage{
load_molecule_info(gbm)
}
\arguments{
\item{gbm}{GeneBCMatrix object to load into}
}
\value{
A GeneBCMatrix object containing the loaded molecule info
}
\description{
Load molecule info h5 file from the Cell Ranger pipeline
}
\examples{
\dontrun{
gbm1 = load_molecule_info(gbm1)
# Note molecule info stored as gbm1@molecule_info
}
}

