% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{run_tsne}
\alias{run_tsne}
\title{Project dimensionally reduced gene-barcode matrix into a visualizable space via t-SNE}
\usage{
run_tsne(pca, ...)
}
\arguments{
\item{pca}{A PCA result from run_pca()}

\item{...}{parameters to pass to Rtsne}
}
\value{
See ?Rtsne in the Rtsne package. The item of interest is the 'Y' element which contains the coordinates.
}
\description{
Project dimensionally reduced gene-barcode matrix into a visualizable space via t-SNE
}
\examples{
\dontrun{
pca <- run_pca(gbm1, n_pcs=10)
tsne_res <- run_tsne(pca)
plot(tsne_res$Y[,1], tsne_res$Y[,2])
}
}

