% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{sparse_pca}
\alias{sparse_pca}
\title{Run a fast PCA algorithm while maintaining sparsity}
\usage{
sparse_pca(x, n_pcs, mu = NULL, s = NULL, center_scale = TRUE)
}
\arguments{
\item{x}{a sparse matrix to run PCA on}

\item{n_pcs}{number of principal components to compute}

\item{mu}{column means}

\item{s}{column standard deviations}

\item{center_scale}{perform centering and scaling}
}
\value{
A list containing \itemize{
\item "x" - The rotated data matrix where rows are barcodes and columns are PCs
\item "sdev" - the standard deviations of the principal components (i.e., sqrt of eigvals of the covariance matrix)
\item "rotation" - The loadings (eigenvectors) where each column is a PC
\item "tot_var" - The total variation in the scaled and centered matrix (this is also the effective rank of the matrix)
\item "var_pcs" - The proportion of variance explained by each principle comoponent
}
}
\description{
Run a fast PCA algorithm while maintaining sparsity
}
\examples{
\dontrun{
pca <- sparse_pca(x, n_pcs=10)
plot(pca$x[,1], pca$x[,2], xlab='PC1', ylab='PC2')
}
}

